+++ nss_mcdb_netdb.c
@@ -599,7 +599,7 @@ nss_mcdb_netdb_gethost_filladdr(const void * const restrict addr,
     hostbuf->h_length    = h_length;
     hostbuf->h_addr_list = (char **)(aligned+8);
     hostbuf->h_aliases[0]   = NULL;
-    hostbuf->h_addr_list[0] = memcpy(aligned+16, addr, (size_t)h_length);
+    hostbuf->h_addr_list[0] = memcpy(aligned+16, addr, (uint32_t)h_length);
     return NSS_STATUS_SUCCESS;
 }
 
@@ -625,10 +625,10 @@ nss_mcdb_netdb_hostent_decode(struct mcdb * const restrict m,
     size_t he_mem_num;
     size_t he_lst_num;
     union { uint32_t u[NSS_HE_HDRSZ>>2]; uint16_t h[NSS_HE_HDRSZ>>1]; } hdr;
-    const uint32_t type = (((unsigned char *)buf)[0]<<24)
-                        | (((unsigned char *)buf)[1]<<16)
-                        | (((unsigned char *)buf)[2]<<8)
-                        |  ((unsigned char *)buf)[3];
+    const uint32_t type = (uint32_t)(((unsigned char *)buf)[0]<<24)
+                                   |(((unsigned char *)buf)[1]<<16)
+                                   |(((unsigned char *)buf)[2]<<8)
+                                   | ((unsigned char *)buf)[3];
 
     /* match type (e.g. AF_INET, AF_INET6), if not zero (AF_UNSPEC) */
     if (type != 0) { /*network byte order; big-endian*/
